fixed CF Icon
