--- admin/src/components/AITextIcon/index.js
@@ -1,25 +1,0 @@
-import React from 'react';
-import styled from 'styled-components';
-import { Icon } from '@strapi/design-system/Icon';
-import { Flex } from '@strapi/design-system/Flex';
-import Alien from '@strapi/icons/Alien';
-
-const IconBox = styled(Flex)`
-  /* Hard code color values */
-  /* to stay consistent between themes */
-  background-color: #f0f0ff; /* primary100 */
-  border: 1px solid #d9d8ff; /* primary200 */
-  svg > path {
-    fill: #4945ff; /* primary600 */
-  }
-`;
-
-const MyIcon = () => {
-  return (
-    <IconBox justifyContent="center" alignItems="center" width={7} height={6} hasRadius aria-hidden>
-      <Icon as={Alien} />
-    </IconBox>
-  );
-};
-
-export default AITextIcon;
