+++ admin/src/components/PluginIcon/index.js
@@ -5,8 +5,28 @@
  */
 
 import React from 'react';
-import Puzzle from '@strapi/icons/Puzzle';
+import styled from 'styled-components';
+import { Icon } from '@strapi/design-system/Icon';
+import { Flex } from '@strapi/design-system/Flex';
+import Pencil from '@strapi/icons/Pencil';
 
-const PluginIcon = () => <Puzzle />;
+
+const IconBox = styled(Flex)`
+  /* Hard code color values */
+  /* to stay consistent between themes */
+  background-color: #f0f0ff; /* primary100 */
+  border: 1px solid #d9d8ff; /* primary200 */
+  svg > path {
+    fill: #4945ff; /* primary600 */
+  }
+`;
+
+const PluginIcon = () => {
+    return (
+        <IconBox justifyContent="center" alignItems="center" width={7} height={6} hasRadius aria-hidden>
+            <Icon as={Pencil} />
+        </IconBox>
+    );
+};
 
 export default PluginIcon;
